Lee Elliott:
Here are some updates to the B-52F.

There are tweaks to the auto-landing system - it seems pretty
reliable now - and I've combined and rearranged the textures so
that it's easier to do re-paints.  I still need to add specific
areas for the engine pods though, as they're currently mapped on
the the fuselage side.  The down side is that the texture is
quite a bit larger.  Was there any consensus on using aggresive
compression for rgb texture files?  It didn't make a huge
difference to the file size but did reduce it a little.  The
texture I've included uses standard, non-aggressive rle
compression.

I've also tweaked the drop view update so that the view position
is interpolated to the new location over a period of five
seconds.  The effect is that the camera flies to the new
location instead of jumping to it.