Lee Elliott:
It still needs a lot of refinement but I think everything's in
the right ball-park.  The lift and drag numbers from the
stand-alone YASim solver are much more realistic and make it
worthwhile.

The auto-takeoff and landing scripts both function but the
landing one needs a lot more work - it floats far too long,
which is something you can't afford to do in a Buff, and then
comes down a bit heavy.

I find these scripts very helpful while I'm working on the YASim
config because they're a lot more consistent than my flying  :)
and they also show the appropriate speeds for doing things i.e
flap retraction/extension while proving that it can be done.

Quite a few of the files have changed, apart from the YASim
config and autopilot, as I fiddled with a few other things as
well as removing my spam-bombed e-mail address, although it's
probably a bit late for that now:(  I made the model more shiny
too, so although the texture and geometry are identical there
needs to be a new .ac file for the new shininess value.