Lee Elliott:
Here's an update for the B-52F.

The most important update is to specify the correct number of
fuel tanks and to size and distribute them more realistically.
At the same time I've squared the elevator control input axis to
try to make it a bit less sensitive to pitch changes.  The
trouble is that there's no real way of simulating the extremely
heavy stick and rudder forces of the b-52 (> 200lbs for the
rudder) which naturally prevents excessive control input.

I've also been experimenting with the auto take-off and landing
scripts and there are a few other, mostly cosmetic, changes e.g.
the aoa indexer now has the correct settings.