Update from Lee Elliott:
Here's an update for the B-52F that attempts to make some
provision for landing in cross-winds.  The rear steering is
automatic - it sets the correct alignment during approach and
then straightens on touch-down out over the course of the next
15 seconds.

I'm not convinced that this is the best way to handle this.
Perhaps I should start straightening the rear gear once the a/c
slowed to below a certain speed i.e. 20 kts or whether I should
add another control and leave it up to the pilot to choose when
to straighten the rear gear.

I had to add two new keyboard controls to handle ground steering
because I needed to separate the ground steering from the rudder
so I used Ctrl-\ & Ctrl-z.

There are a couple of files in cvs that can be removed - these
are:

B-52F/Models/B52-F-Side-01.rgb
B-52F/Models/B52-F-Top-01.rgb


Then a second update:

I've attached an update for the B-52F to address the broken
panels and instruments caused by the clipping patch.

I'm afraid it'll be a while before I get around to fixing all of
the aircraft I've done because each affected instrument and
panel needs to be re-sized and re-aligned, which takes a lot of
time.

This update should be applied after the steering update.